moved javascript code from browser-overlay.xul to browser-overlay.js (refactoring?)
Reason of change: Andres Hernandez's comment
> Please move the code on the browser-overlay.xul file into a js file wrap inside a namespace, in order to prevent conflicts with other add-ons that may be installed by users. ... Fix it for your next update.

The older browser-overlay.xul code does not seems to occur any conflicts (because script on "oncommand" is running on a closure). But he said "Fix it for your next update." So the code must be replaced. It is not a fix but a refactoring, I think.